TrainBackends: Make sure the output of load is a model
